Task: Refactor code into appropriate directories
Task: Refactor code into appropriate directories
