+++ CsvReader/CSVReader.py
@@ -1,5 +1,5 @@
 import csv
-from absolutepath import absolutepath
+from src.absolutepath import absolutepath
 
 
 def ClassFactory(class_name, dictionary):
